+++ alt-text/src/exports/server.ts
@@ -1,0 +1,8 @@
+export { AltTextHealthWidget } from '../components/AltTextHealthWidget.js'
+export { getAltTextHealth } from '../utilities/altTextHealth.js'
+export type {
+  AltTextHealthError,
+  AltTextHealthErrorCode,
+  AltTextHealthScan,
+  AltTextHealthScanCollection,
+} from '../utilities/altTextHealth.js'